Improved the image area a great deal
Replaced the static demo image with a graphics area, and added a menubar that I can bind to the files in a loaded folder. Then the user can just click the file they want to work on and it will open it. (some kinks may need to be worked out since this could mean loading lots of files at once...)